[kda合入] 合入kda的op/l2norm，删除_kda_utils
