+++ fla/ops/triton/triton_core/kda/chunk.py
@@ -8,7 +8,7 @@
 
 import torch
 
-from fla.ops.triton.triton_core.kda._kda_utils.l2norm import l2norm_bwd, l2norm_fwd
+from fla.ops.triton.triton_core.l2norm import l2norm_bwd, l2norm_fwd
 from fla.ops.triton.triton_core.kda._kda_cp import FLACPContext
 from fla.ops.triton.triton_core.kda.chunk_bwd import chunk_kda_bwd
 from fla.ops.triton.triton_core.kda.chunk_fwd import chunk_kda_fwd
