+++ fla/ops/triton/triton_core/kda/chunk_intra_token_parallel.py
@@ -10,7 +10,7 @@ import torch
 import triton
 import triton.language as tl
 
-from fla.ops.triton.triton_core.kda._kda_utils.op import exp2
+from fla.ops.triton.triton_core.utils import exp2
 from fla.ops.triton.triton_core.utils import autotune_cache_kwargs
 
 @triton.heuristics({
