+++ fla/ops/triton/triton_core/kda/fused_recurrent.py
@@ -10,7 +10,7 @@ import torch
 import triton
 import triton.language as tl
 
-from fla.ops.triton.triton_core.kda._kda_utils.op import exp
+from fla.ops.triton.triton_core.utils import exp
 from fla.ops.triton.triton_core.softplus import softplus
 from fla.ops.triton.triton_core.utils import input_guard
