+++ fla/ops/triton/triton_core/kda/gate.py
@@ -12,7 +12,7 @@ import triton
 import triton.language as tl
 
 from fla.ops.triton.triton_core.index import prepare_chunk_indices
-from fla.ops.triton.triton_core.kda._kda_utils.op import exp
+from fla.ops.triton.triton_core.utils import exp
 from fla.ops.triton.triton_core.softplus import softplus
 from fla.ops.triton.triton_core.utils import autocast_custom_bwd, autocast_custom_fwd, autotune_cache_kwargs, input_guard
