+++ fla/ops/triton/triton_core/kda/wy_fast.py
@@ -11,7 +11,7 @@ import triton
 import triton.language as tl
 
 from fla.ops.triton.triton_core.index import prepare_chunk_indices
-from fla.ops.triton.triton_core.kda._kda_utils.op import exp2
+from fla.ops.triton.triton_core.utils import exp2
 from fla.ops.triton.triton_core.utils import autotune_cache_kwargs, check_shared_mem
 
 @triton.heuristics({
